Move Rosetta integration tests to correct files
